move DefaultPartitionGrouper
