+++ streams/src/test/java/org/apache/kafka/streams/processor/internals/SingleGroupPartitionGrouperStub.java
@@ -30,7 +30,7 @@ import java.util.Set;
 @SuppressWarnings("deprecation")
 public class SingleGroupPartitionGrouperStub implements org.apache.kafka.streams.processor.PartitionGrouper {
     private org.apache.kafka.streams.processor.PartitionGrouper defaultPartitionGrouper =
-        new org.apache.kafka.streams.processor.DefaultPartitionGrouper();
+        new DefaultPartitionGrouper();
 
     @Override
     public Map<TaskId, Set<TopicPartition>> partitionGroups(final Map<Integer, Set<String>> topicGroups, final Cluster metadata) {
